Move ride call to actions
