+++ apps/dashboard/src/app/page.tsx
@@ -4,27 +4,11 @@ import { prisma } from 'db'
 import { PageLayout } from '@components/page-layout'
 import { DashboardTable } from '@components/DashboardTable'
 import { AddRideForm } from '@components/AddRideForm'
-import { addRide, addPassenger } from '@actions/ride'
+import { addRide, addPassenger, getAllRides } from '@actions/ride'
 
 export const metadata: Metadata = {
   title: 'Web - Turborepo Example',
 }
-
-const getAllRides = async () => {
-  const rides = await prisma.ride.findMany({
-    include: {
-      driver: true,
-      passenger: true,
-    },
-    take: 10, // for now, get all of them later
-    orderBy: {
-      createdAt: 'desc',
-    },
-  })
-  return rides
-}
-
-export type InferredRide = Awaited<ReturnType<typeof getAllRides>>[0]
 
 export default async function Home() {
   const rides = await getAllRides()
